Improve logging of session

+++ src/app/socks5/traffic.rs
@@ -1,9 +1,9 @@
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 
 #[derive(Default, Debug)]
 pub(crate) struct Usage {
-    tx_bytes: AtomicUsize,
-    rx_bytes: AtomicUsize,
+    tx_bytes: AtomicU64,
+    rx_bytes: AtomicU64,
     session_active: AtomicUsize,
     session_total: AtomicUsize,
 }
@@ -20,11 +20,11 @@ impl Usage {
 
     #[inline]
     pub(super) fn add_tx(&self, bytes: usize) {
-        self.tx_bytes.fetch_add(bytes, Ordering::Relaxed);
+        self.tx_bytes.fetch_add(bytes as u64, Ordering::Relaxed);
     }
 
     #[inline]
     pub(super) fn add_rx(&self, bytes: usize) {
-        self.rx_bytes.fetch_add(bytes, Ordering::Relaxed);
+        self.rx_bytes.fetch_add(bytes as u64, Ordering::Relaxed);
     }
 }
